Remove compiler warnings.
git-svn-id: https://ggt.svn.sourceforge.net/svnroot/ggt/trunk@950 6382ae23-ed94-47d5-a97d-470f0bc4535f

+++ GGT/modules/GMTL/Test/TestSuite/TestCases/VecGenTest.cpp
@@ -7,8 +7,8 @@
  *
  * -----------------------------------------------------------------
  * File:          $RCSfile: VecGenTest.cpp,v $
- * Date modified: $Date: 2003-09-24 00:03:18 $
- * Version:       $Revision: 1.3 $
+ * Date modified: $Date: 2004-08-19 16:50:40 $
+ * Version:       $Revision: 1.4 $
  * -----------------------------------------------------------------
  *
  *********************************************************** ggt-head end */
@@ -58,7 +58,7 @@ namespace gmtlTest
       CPPUNIT_ASSERT( vec[1] == 21.0f );
       CPPUNIT_ASSERT( vec[2] == 31.0f );
    }
-   
+
    void VecGenTest::testMakeNormalVecFromVec()
    {
       {
@@ -67,14 +67,14 @@ namespace gmtlTest
          gmtl::Vec<float, 3> result( gmtl::makeNormal( vec ) );
          CPPUNIT_ASSERT( gmtl::isEqual( expected, result, eps ) );
       }
-      
+
       {
          const float eps = 0.001f;
          const gmtl::Vec<float, 3> vec( 0.1f, 128943139.0f, 0.0f ), expected( 0, 1, 0 );
          gmtl::Vec<float, 3> result( gmtl::makeNormal( vec ) );
          CPPUNIT_ASSERT( gmtl::isEqual( expected, result, eps ) );
       }
-      
+
       {
          const float eps = 0.001f;
          const gmtl::Vec<float, 3> vec( 0.0f, 0.0f, 0.001f ), expected( 0, 0, 1 );
@@ -141,17 +141,17 @@ namespace gmtlTest
       {
          // Create a matrix filled in linearly
          gmtl::Matrix<T, ROWS, COLS> mat;
-         for (unsigned int i=0; i<(ROWS*COLS); ++i)
+         for (unsigned i=0; i<(unsigned)(ROWS*COLS); ++i)
          {
             mat.mData[i] = T(i);
          }
 
          // Get each column and test it
-         for (unsigned int i=0; i<COLS; ++i)
+         for (unsigned i=0; i<unsigned(COLS); ++i)
          {
             gmtl::Vec<T, ROWS> col = gmtl::makeColumn(mat, i);
             // Test the row values
-            for (unsigned int row=0; row<ROWS; ++row)
+            for (unsigned row=0; row<unsigned(ROWS); ++row)
             {
                CPPUNIT_ASSERT(col[row] == (i*ROWS + row));
             }
@@ -185,8 +185,8 @@ namespace gmtlTest
 
 
    //-- timing tests --//
-   
-   
+
+
    void VecGenMetricTest::testGenTimingMakeNormalVec1()
    {
       gmtl::Vec<double, 4> vec4d;
@@ -199,11 +199,11 @@ namespace gmtlTest
       CPPUNIT_METRIC_STOP_TIMING();
       CPPUNIT_ASSERT_METRIC_TIMING_LE( "VecGenTest/makeNormal(vec4d)", iters, 0.075f, 0.1f);  // warn at 7.5%, error at 10%
 
-      // force intelligent compilers to do all the iterations (ie. to not optimize them out), 
+      // force intelligent compilers to do all the iterations (ie. to not optimize them out),
       // by using the variables computed...
       CPPUNIT_ASSERT( vec4d[0] != 10000.0f && vec4d[1] != 10000.0f && vec4d[2] != 10000.0f );
    }
-   
+
    void VecGenMetricTest::testGenTimingMakeNormalVec2()
    {
       gmtl::Vec<float, 4> vec4f;
@@ -215,8 +215,8 @@ namespace gmtlTest
       }
       CPPUNIT_METRIC_STOP_TIMING();
       CPPUNIT_ASSERT_METRIC_TIMING_LE( "VecGenTest/makeNormal(vec4f)", iters, 0.075f, 0.1f);  // warn at 7.5%, error at 10%
-   
-      // force intelligent compilers to do all the iterations (ie. to not optimize them out), 
+
+      // force intelligent compilers to do all the iterations (ie. to not optimize them out),
       // by using the variables computed...
       CPPUNIT_ASSERT( vec4f[0] != 10000.0f && vec4f[1] != 10000.0f && vec4f[2] != 10000.0f );
    }
@@ -232,11 +232,11 @@ namespace gmtlTest
       CPPUNIT_METRIC_STOP_TIMING();
       CPPUNIT_ASSERT_METRIC_TIMING_LE( "VecGenTest/makeNormal(vec3d)", iters, 0.075f, 0.1f);  // warn at 7.5%, error at 10%
 
-      // force intelligent compilers to do all the iterations (ie. to not optimize them out), 
+      // force intelligent compilers to do all the iterations (ie. to not optimize them out),
       // by using the variables computed...
       CPPUNIT_ASSERT( vec3d[0] != 10000.0f && vec3d[1] != 10000.0f && vec3d[2] != 10000.0f );
    }
-   
+
    void VecGenMetricTest::testGenTimingMakeNormalVec4()
    {
       gmtl::Vec<float, 3> vec3f;
@@ -248,20 +248,20 @@ namespace gmtlTest
       }
       CPPUNIT_METRIC_STOP_TIMING();
       CPPUNIT_ASSERT_METRIC_TIMING_LE( "VecGenTest/makeNormal(vec3f)", iters, 0.075f, 0.1f);  // warn at 7.5%, error at 10%
-   
-      
-      // force intelligent compilers to do all the iterations (ie. to not optimize them out), 
+
+
+      // force intelligent compilers to do all the iterations (ie. to not optimize them out),
       // by using the variables computed...
       CPPUNIT_ASSERT( vec3f[0] != 10000.0f && vec3f[1] != 10000.0f && vec3f[2] != 10000.0f );
    }
-   
+
    void VecGenMetricTest::testGenTimingMakeVec()
    {
       gmtl::Quat<double> q1;
       gmtl::Vec<double, 3> v1;
       const long iters(25000);
       CPPUNIT_METRIC_START_TIMING();
-      
+
       for (long iter = 0; iter < iters; ++iter)
       {
          v1 = gmtl::makeVec( q1 );
@@ -278,8 +278,8 @@ namespace gmtlTest
       }
       CPPUNIT_METRIC_STOP_TIMING();
       CPPUNIT_ASSERT_METRIC_TIMING_LE( "VecGenTest/makeVec(quatf)", iters, 0.075f, 0.1f);  // warn at 7.5%, error at 10%
-   
-      // force intelligent compilers to do all the iterations (ie. to not optimize them out), 
+
+      // force intelligent compilers to do all the iterations (ie. to not optimize them out),
       // by using the variables computed...
       CPPUNIT_ASSERT( v2[0] != 10000.0f );
       CPPUNIT_ASSERT( v1[0] != 10000.0f );
